Module aws_sdk_organizations::error

Modules

access_denied_exception

See AccessDeniedException

access_denied_for_dependency_exception

See AccessDeniedForDependencyException

account_already_registered_exception

See AccountAlreadyRegisteredException

account_not_found_exception

See AccountNotFoundException

account_not_registered_exception

See AccountNotRegisteredException

account_owner_not_verified_exception

See AccountOwnerNotVerifiedException

already_in_organization_exception

See AlreadyInOrganizationException

aws_organizations_not_in_use_exception

See AwsOrganizationsNotInUseException

child_not_found_exception

See ChildNotFoundException

concurrent_modification_exception

See ConcurrentModificationException

constraint_violation_exception

See ConstraintViolationException

create_account_status_not_found_exception

See CreateAccountStatusNotFoundException

destination_parent_not_found_exception

See DestinationParentNotFoundException

duplicate_account_exception

See DuplicateAccountException

duplicate_handshake_exception

See DuplicateHandshakeException

duplicate_organizational_unit_exception

See DuplicateOrganizationalUnitException

duplicate_policy_attachment_exception

See DuplicatePolicyAttachmentException

duplicate_policy_exception

See DuplicatePolicyException

effective_policy_not_found_exception

See EffectivePolicyNotFoundException

finalizing_organization_exception

See FinalizingOrganizationException

handshake_already_in_state_exception

See HandshakeAlreadyInStateException

handshake_constraint_violation_exception

See HandshakeConstraintViolationException

handshake_not_found_exception

See HandshakeNotFoundException

invalid_handshake_transition_exception

See InvalidHandshakeTransitionException

invalid_input_exception

See InvalidInputException

malformed_policy_document_exception

See MalformedPolicyDocumentException

master_cannot_leave_organization_exception

See MasterCannotLeaveOrganizationException

organization_not_empty_exception

See OrganizationNotEmptyException

organizational_unit_not_empty_exception

See OrganizationalUnitNotEmptyException

organizational_unit_not_found_exception

See OrganizationalUnitNotFoundException

parent_not_found_exception

See ParentNotFoundException

policy_changes_in_progress_exception

See PolicyChangesInProgressException

policy_in_use_exception

See PolicyInUseException

policy_not_attached_exception

See PolicyNotAttachedException

policy_not_found_exception

See PolicyNotFoundException

policy_type_already_enabled_exception

See PolicyTypeAlreadyEnabledException

policy_type_not_available_for_organization_exception

See PolicyTypeNotAvailableForOrganizationException

policy_type_not_enabled_exception

See PolicyTypeNotEnabledException

root_not_found_exception

See RootNotFoundException

service_exception

See ServiceException

source_parent_not_found_exception

See SourceParentNotFoundException

target_not_found_exception

See TargetNotFoundException

too_many_requests_exception

See TooManyRequestsException

unsupported_api_endpoint_exception

See UnsupportedApiEndpointException

Structs

AcceptHandshakeError

AccessDeniedException

You don't have permissions to perform the requested operation. The user or role that

AccessDeniedForDependencyException

The operation that you attempted requires you to have the

AccountAlreadyRegisteredException

The specified account is already a delegated administrator for this AWS

AccountNotFoundException

We can't find an AWS account with the AccountId that you specified, or

AccountNotRegisteredException

The specified account is not a delegated administrator for this AWS service.

AccountOwnerNotVerifiedException

You can't invite an existing account to your organization until you verify that you

AlreadyInOrganizationException

This account is already a member of an organization. An account can belong to only one

AttachPolicyError

AwsOrganizationsNotInUseException

Your account isn't a member of an organization. To make this request, you must use the

CancelHandshakeError

ChildNotFoundException

We can't find an organizational unit (OU) or AWS account with the

ConcurrentModificationException

The target of the operation is currently being modified by a different request. Try

ConstraintViolationException

Performing this operation violates a minimum or maximum value limit. For example,

CreateAccountError

CreateAccountStatusNotFoundException

We can't find an create account request with the CreateAccountRequestId

CreateGovCloudAccountError

CreateOrganizationError

CreateOrganizationalUnitError

CreatePolicyError

DeclineHandshakeError

DeleteOrganizationError

DeleteOrganizationalUnitError

DeletePolicyError

DeregisterDelegatedAdministratorError

DescribeAccountError

DescribeCreateAccountStatusError

DescribeEffectivePolicyError

DescribeHandshakeError

DescribeOrganizationError

DescribeOrganizationalUnitError

DescribePolicyError

DestinationParentNotFoundException

We can't find the destination container (a root or OU) with the ParentId

DetachPolicyError

DisableAWSServiceAccessError

DisablePolicyTypeError

DuplicateAccountException

That account is already present in the specified destination.

DuplicateHandshakeException

A handshake with the same action and target already exists. For example, if you

DuplicateOrganizationalUnitException

An OU with the same name already exists.

DuplicatePolicyAttachmentException

The selected policy is already attached to the specified target.

DuplicatePolicyException

A policy with the same name already exists.

EffectivePolicyNotFoundException

If you ran this action on the management account, this policy type is not enabled. If

EnableAWSServiceAccessError

EnableAllFeaturesError

EnablePolicyTypeError

FinalizingOrganizationException

AWS Organizations couldn't perform the operation because your organization hasn't finished

HandshakeAlreadyInStateException

The specified handshake is already in the requested state. For example, you can't

HandshakeConstraintViolationException

The requested operation would violate the constraint identified in the reason

HandshakeNotFoundException

We can't find a handshake with the HandshakeId that you specified.

InvalidHandshakeTransitionException

You can't perform the operation on the handshake in its current state. For example,

InvalidInputException

The requested operation failed because you provided invalid values for one or more of

InviteAccountToOrganizationError

LeaveOrganizationError

ListAWSServiceAccessForOrganizationError

ListAccountsError

ListAccountsForParentError

ListChildrenError

ListCreateAccountStatusError

ListDelegatedAdministratorsError

ListDelegatedServicesForAccountError

ListHandshakesForAccountError

ListHandshakesForOrganizationError

ListOrganizationalUnitsForParentError

ListParentsError

ListPoliciesError

ListPoliciesForTargetError

ListRootsError

ListTagsForResourceError

ListTargetsForPolicyError

MalformedPolicyDocumentException

The provided policy document doesn't meet the requirements of the specified policy

MasterCannotLeaveOrganizationException

You can't remove a management account from an organization. If you want the management

MoveAccountError

OrganizationNotEmptyException

The organization isn't empty. To delete an organization, you must first remove all

OrganizationalUnitNotEmptyException

The specified OU is not empty. Move all accounts to another root or to other OUs,

OrganizationalUnitNotFoundException

We can't find an OU with the OrganizationalUnitId that you

ParentNotFoundException

We can't find a root or OU with the ParentId that you specified.

PolicyChangesInProgressException

Changes to the effective policy are in progress, and its contents can't be returned.

PolicyInUseException

The policy is attached to one or more entities. You must detach it from all roots,

PolicyNotAttachedException

The policy isn't attached to the specified target in the specified root.

PolicyNotFoundException

We can't find a policy with the PolicyId that you specified.

PolicyTypeAlreadyEnabledException

The specified policy type is already enabled in the specified root.

PolicyTypeNotAvailableForOrganizationException

You can't use the specified policy type with the feature set currently enabled for

PolicyTypeNotEnabledException

The specified policy type isn't currently enabled in this root. You can't attach

RegisterDelegatedAdministratorError

RemoveAccountFromOrganizationError

RootNotFoundException

We can't find a root with the RootId that you specified.

ServiceException

AWS Organizations can't complete your request because of an internal service error. Try again

SourceParentNotFoundException

We can't find a source root or OU with the ParentId that you

TagResourceError

TargetNotFoundException

We can't find a root, OU, account, or policy with the TargetId that you

TooManyRequestsException

You have sent too many requests in too short a period of time. The quota helps protect

UnsupportedApiEndpointException

This action isn't available in the current AWS Region.

UntagResourceError

UpdateOrganizationalUnitError

UpdatePolicyError

Enums

AcceptHandshakeErrorKind

AttachPolicyErrorKind

CancelHandshakeErrorKind

CreateAccountErrorKind

CreateGovCloudAccountErrorKind

CreateOrganizationErrorKind

CreateOrganizationalUnitErrorKind

CreatePolicyErrorKind

DeclineHandshakeErrorKind

DeleteOrganizationErrorKind

DeleteOrganizationalUnitErrorKind

DeletePolicyErrorKind

DeregisterDelegatedAdministratorErrorKind

DescribeAccountErrorKind

DescribeCreateAccountStatusErrorKind

DescribeEffectivePolicyErrorKind

DescribeHandshakeErrorKind

DescribeOrganizationErrorKind

DescribeOrganizationalUnitErrorKind

DescribePolicyErrorKind

DetachPolicyErrorKind

DisableAWSServiceAccessErrorKind

DisablePolicyTypeErrorKind

EnableAWSServiceAccessErrorKind

EnableAllFeaturesErrorKind

EnablePolicyTypeErrorKind

InviteAccountToOrganizationErrorKind

LeaveOrganizationErrorKind

ListAWSServiceAccessForOrganizationErrorKind

ListAccountsErrorKind

ListAccountsForParentErrorKind

ListChildrenErrorKind

ListCreateAccountStatusErrorKind

ListDelegatedAdministratorsErrorKind

ListDelegatedServicesForAccountErrorKind

ListHandshakesForAccountErrorKind

ListHandshakesForOrganizationErrorKind

ListOrganizationalUnitsForParentErrorKind

ListParentsErrorKind

ListPoliciesErrorKind

ListPoliciesForTargetErrorKind

ListRootsErrorKind

ListTagsForResourceErrorKind

ListTargetsForPolicyErrorKind

MoveAccountErrorKind

RegisterDelegatedAdministratorErrorKind

RemoveAccountFromOrganizationErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateOrganizationalUnitErrorKind

UpdatePolicyErrorKind