Enum aws_sdk_organizations::error::CancelHandshakeErrorKind

#[non_exhaustive]
pub enum CancelHandshakeErrorKind {
    AccessDeniedException(AccessDeniedException),
    ConcurrentModificationException(ConcurrentModificationException),
    HandshakeAlreadyInStateException(HandshakeAlreadyInStateException),
    HandshakeNotFoundException(HandshakeNotFoundException),
    InvalidHandshakeTransitionException(InvalidHandshakeTransitionException),
    InvalidInputException(InvalidInputException),
    ServiceException(ServiceException),
    TooManyRequestsException(TooManyRequestsException),
    Unhandled(Box<dyn Error + Send + Sync + 'static>),
}

Types of errors that can occur for the CancelHandshake operation.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

AccessDeniedException(AccessDeniedException)

You don't have permissions to perform the requested operation. The user or role that is making the request must have at least one IAM permissions policy attached that grants the required permissions. For more information, see Access Management in the IAM User Guide.

Tuple Fields of AccessDeniedException

0: AccessDeniedException

ConcurrentModificationException(ConcurrentModificationException)

The target of the operation is currently being modified by a different request. Try again later.

Tuple Fields of ConcurrentModificationException

0: ConcurrentModificationException

HandshakeAlreadyInStateException(HandshakeAlreadyInStateException)

The specified handshake is already in the requested state. For example, you can't accept a handshake that was already accepted.

Tuple Fields of HandshakeAlreadyInStateException

0: HandshakeAlreadyInStateException

HandshakeNotFoundException(HandshakeNotFoundException)

We can't find a handshake with the HandshakeId that you specified.

Tuple Fields of HandshakeNotFoundException

0: HandshakeNotFoundException

InvalidHandshakeTransitionException(InvalidHandshakeTransitionException)

You can't perform the operation on the handshake in its current state. For example, you can't cancel a handshake that was already accepted or accept a handshake that was already declined.

Tuple Fields of InvalidHandshakeTransitionException

0: InvalidHandshakeTransitionException

InvalidInputException(InvalidInputException)

The requested operation failed because you provided invalid values for one or more of the request parameters. This exception includes a reason that contains additional information about the violated limit:

Some of the reasons in the following list might not be applicable to this specific API or operation.

DUPLICATE_TAG_KEY: Tag keys must be unique among the tags attached to the same entity.

IMMUTABLE_POLICY: You specified a policy that is managed by AWS and can't be modified.

INPUT_REQUIRED: You must include a value for all required parameters.

INVALID_EMAIL_ADDRESS_TARGET: You specified an invalid email address for the invited account owner.

INVALID_ENUM: You specified an invalid value.

INVALID_ENUM_POLICY_TYPE: You specified an invalid policy type string.

INVALID_FULL_NAME_TARGET: You specified a full name that contains invalid characters.

INVALID_LIST_MEMBER: You provided a list to a parameter that contains at least one invalid value.

INVALID_PAGINATION_TOKEN: Get the value for the NextToken parameter from the response to a previous call of the operation.

INVALID_PARTY_TYPE_TARGET: You specified the wrong type of entity (account, organization, or email) as a party.

INVALID_PATTERN: You provided a value that doesn't match the required pattern.

INVALID_PATTERN_TARGET_ID: You specified a policy target ID that doesn't match the required pattern.

INVALID_ROLE_NAME: You provided a role name that isn't valid. A role name can't begin with the reserved prefix AWSServiceRoleFor.

INVALID_SYNTAX_ORGANIZATION_ARN: You specified an invalid Amazon Resource Name (ARN) for the organization.

INVALID_SYNTAX_POLICY_ID: You specified an invalid policy ID.

INVALID_SYSTEM_TAGS_PARAMETER: You specified a tag key that is a system tag. You can’t add, edit, or delete system tag keys because they're reserved for AWS use. System tags don’t count against your tags per resource limit.

MAX_FILTER_LIMIT_EXCEEDED: You can specify only one filter parameter for the operation.

MAX_LENGTH_EXCEEDED: You provided a string parameter that is longer than allowed.

MAX_VALUE_EXCEEDED: You provided a numeric parameter that has a larger value than allowed.

MIN_LENGTH_EXCEEDED: You provided a string parameter that is shorter than allowed.

MIN_VALUE_EXCEEDED: You provided a numeric parameter that has a smaller value than allowed.

MOVING_ACCOUNT_BETWEEN_DIFFERENT_ROOTS: You can move an account only between entities in the same root.

TARGET_NOT_SUPPORTED: You can't perform the specified operation on that target entity.

UNRECOGNIZED_SERVICE_PRINCIPAL: You specified a service principal that isn't recognized.

Tuple Fields of InvalidInputException

0: InvalidInputException

ServiceException(ServiceException)

AWS Organizations can't complete your request because of an internal service error. Try again later.

Tuple Fields of ServiceException

0: ServiceException

TooManyRequestsException(TooManyRequestsException)

You have sent too many requests in too short a period of time. The quota helps protect against denial-of-service attacks. Try again later.

For information about quotas that affect AWS Organizations, see Quotas for AWS Organizationsin the AWS Organizations User Guide.

Tuple Fields of TooManyRequestsException

0: TooManyRequestsException

Unhandled(Box<dyn Error + Send + Sync + 'static>)

An unexpected error, e.g. invalid JSON returned by the service or an unknown error code

Tuple Fields of Unhandled

0: Box<dyn Error + Send + Sync + 'static>