Module aws_sdk_organizations::model

Modules

account

See Account

child

See Child

create_account_status

See CreateAccountStatus

delegated_administrator

See DelegatedAdministrator

delegated_service

See DelegatedService

effective_policy

See EffectivePolicy

enabled_service_principal

See EnabledServicePrincipal

handshake

See Handshake

handshake_filter

See HandshakeFilter

handshake_party

See HandshakeParty

handshake_resource

See HandshakeResource

organization

See Organization

organizational_unit

See OrganizationalUnit

parent

See Parent

policy

See Policy

policy_summary

See PolicySummary

policy_target_summary

See PolicyTargetSummary

policy_type_summary

See PolicyTypeSummary

root

See Root

tag

See Tag

Structs

Account

Contains information about an AWS account that is a member of an

Child

Contains a list of child entities, either OUs or accounts.

CreateAccountStatus

Contains the status about a CreateAccount or CreateGovCloudAccount request to create an AWS account or an AWS

DelegatedAdministrator

Contains information about the delegated administrator.

DelegatedService

Contains information about the AWS service for which the account is a delegated

EffectivePolicy

Contains rules to be applied to the affected accounts. The effective policy is the

EnabledServicePrincipal

A structure that contains details of a service principal that represents an AWS

Handshake

Contains information that must be exchanged to securely establish a relationship

HandshakeFilter

Specifies the criteria that are used to select the handshakes for the

HandshakeParty

Identifies a participant in a handshake.

HandshakeResource

Contains additional data that is needed to process a handshake.

Organization

Contains details about an organization. An organization is a collection of accounts

OrganizationalUnit

Contains details about an organizational unit (OU). An OU is a container of AWS

Parent

Contains information about either a root or an organizational unit (OU) that can

Policy

Contains rules to be applied to the affected accounts. Policies can be attached

PolicySummary

Contains information about a policy, but does not include the content. To see the

PolicyTargetSummary

Contains information about a root, OU, or account that a policy is attached to.

PolicyTypeSummary

Contains information about a policy type and its status in the associated root.

Root

Contains details about a root. A root is a top-level parent node in the hierarchy of

Tag

A custom key-value pair associated with a resource within your organization.

Enums

AccessDeniedForDependencyExceptionReason

AccountJoinedMethod

AccountStatus

ActionType

ChildType

ConstraintViolationExceptionReason

CreateAccountFailureReason

CreateAccountState

EffectivePolicyType

HandshakeConstraintViolationExceptionReason

HandshakePartyType

HandshakeResourceType

HandshakeState

IamUserAccessToBilling

InvalidInputExceptionReason

OrganizationFeatureSet

ParentType

PolicyType

PolicyTypeStatus

TargetType