Module aws_sdk_organizations::operation

Structs

AcceptHandshake

Sends a response to the originator of a handshake agreeing to the action proposed by

AttachPolicy

Attaches a policy to a root, an organizational unit (OU), or an individual account.

CancelHandshake

Cancels a handshake. Canceling a handshake sets the handshake state to

CreateAccount

Creates an AWS account that is automatically a member of the organization whose

CreateGovCloudAccount

This action is available if all of the following are true:

CreateOrganization

Creates an AWS organization. The account whose user is calling the

CreateOrganizationalUnit

Creates an organizational unit (OU) within a root or parent OU. An OU is a container

CreatePolicy

Creates a policy of a specified type that you can attach to a root, an organizational

DeclineHandshake

Declines a handshake request. This sets the handshake state to DECLINED

DeleteOrganization

Deletes the organization. You can delete an organization only by using credentials

DeleteOrganizationalUnit

Deletes an organizational unit (OU) from a root or another OU. You must first remove

DeletePolicy

Deletes the specified policy from your organization. Before you perform this

DeregisterDelegatedAdministrator

Removes the specified member AWS account as a delegated administrator for the

DescribeAccount

Retrieves AWS Organizations-related information about the specified account.

DescribeCreateAccountStatus

Retrieves the current status of an asynchronous request to create an account.

DescribeEffectivePolicy

Returns the contents of the effective policy for specified policy type and account.

DescribeHandshake

Retrieves information about a previously requested handshake. The handshake ID comes

DescribeOrganization

Retrieves information about the organization that the user's account belongs

DescribeOrganizationalUnit

Retrieves information about an organizational unit (OU).

DescribePolicy

Retrieves information about a policy.

DetachPolicy

Detaches a policy from a target root, organizational unit (OU), or account.

DisableAWSServiceAccess

Disables the integration of an AWS service (the service that is specified by

DisablePolicyType

Disables an organizational policy type in a root. A policy of a certain type can be

EnableAWSServiceAccess

Enables the integration of an AWS service (the service that is specified by

EnableAllFeatures

Enables all features in an organization. This enables the use of organization policies

EnablePolicyType

Enables a policy type in a root. After you enable a policy type in a root, you can

InviteAccountToOrganization

Sends an invitation to another account to join your organization as a member account.

LeaveOrganization

Removes a member account from its parent organization. This version of the operation

ListAWSServiceAccessForOrganization

Returns a list of the AWS services that you enabled to integrate with your

ListAccounts

Lists all the accounts in the organization. To request only the accounts in a

ListAccountsForParent

Lists the accounts in an organization that are contained by the specified target root

ListChildren

Lists all of the organizational units (OUs) or accounts that are contained in the

ListCreateAccountStatus

Lists the account creation requests that match the specified status that is currently

ListDelegatedAdministrators

Lists the AWS accounts that are designated as delegated administrators in this

ListDelegatedServicesForAccount

List the AWS services for which the specified account is a delegated

ListHandshakesForAccount

Lists the current handshakes that are associated with the account of the requesting

ListHandshakesForOrganization

Lists the handshakes that are associated with the organization that the requesting

ListOrganizationalUnitsForParent

Lists the organizational units (OUs) in a parent organizational unit or root.

ListParents

Lists the root or organizational units (OUs) that serve as the immediate parent of the

ListPolicies

Retrieves the list of all policies in an organization of a specified type.

ListPoliciesForTarget

Lists the policies that are directly attached to the specified target root,

ListRoots

Lists the roots that are defined in the current organization.

ListTagsForResource

Lists tags that are attached to the specified resource.

ListTargetsForPolicy

Lists all the roots, organizational units (OUs), and accounts that the specified

MoveAccount

Moves an account from its current source parent root or organizational unit (OU) to

RegisterDelegatedAdministrator

Enables the specified member account to administer the Organizations features of the specified

RemoveAccountFromOrganization

Removes the specified account from the organization.

TagResource

Adds one or more tags to the specified resource.

UntagResource

Removes any tags with the specified keys from the specified resource.

UpdateOrganizationalUnit

Renames the specified organizational unit (OU). The ID and ARN don't change. The child

UpdatePolicy

Updates an existing policy with a new name, description, or content. If you don't